+++ stylelint-plugins/rules/css-global/test.js
@@ -9,42 +9,6 @@ test('css-global', {
         filename: 'src/cssGlobal/abc.css',
         code: `:root { --x: 0; }`,
       },
-      {
-        title: 'valid - allow @value in cssGlobal',
-        filename: 'src/cssGlobal/abc.css',
-        code: `
-          @value _x x;
-          :root { --x: 0; }
-        `,
-      },
-      {
-        title: 'valid - allow :export in cssGlobal',
-        filename: 'src/cssGlobal/abc.css',
-        code: `
-          :export { x: 0; }
-          :root { --x: 0; }
-        `
-      },
-      {
-        title: 'valid - allow custom media in globalCss',
-        filename: 'src/cssGlobal/abc.css',
-        code: `@custom-media --x (max-width: 30em);`
-      },
-      {
-        title: 'valid - allow custom properties in globalCss',
-        filename: 'src/cssGlobal/abc.css',
-        code: `:root { --x: 0; }`
-      },
-      {
-        title: 'valid - allow custom selectors in globalCss',
-        filename: 'src/cssGlobal/abc.css',
-        code: `@custom-selector :--x x;`
-      },
-      {
-        title: 'valid - allow @value and :export in other files',
-        filename: 'src/notCssGlobal/abc.css',
-        code: `@value x: x; :export { x: 0; }`
-      },
     ],
     invalid: [
       {
@@ -60,7 +24,7 @@ test('css-global', {
         ],
       },
       {
-        title: 'invalid - only allow @custom-media and @custom-selector in cssGlobal directory',
+        title: 'invalid - only allow :root at-rules in cssGlobal directory',
         filename: 'src/cssGlobal/abc.css',
         code: `
           @keyframes x { }
@@ -74,14 +38,6 @@ test('css-global', {
       {
         code: `:root { --x: 0; }`,
         warnings: [messages['no'](':root')]
-      },
-      {
-        code: `@custom-media --x (max-width: 30em);`,
-        warnings: [messages['no']('@custom-media')]
-      },
-      {
-        code: `@custom-selector :--x x;`,
-        warnings: [messages['no']('@custom-selector')]
       },
     ],
   },