+++ stylelint-plugins/rules/index/test.js
@@ -56,11 +56,6 @@ test('index', {
         filename: 'index.css',
         code: `@import url('https://fonts.googleapis.com/css');`
       },
-      {
-        title: 'allow custom element @kaliber-scoped in index.css',
-        filename: 'index.css',
-        code: `@kaliber-scoped custom-element;`
-      },
     ],
     invalid: [
       {
@@ -68,12 +63,6 @@ test('index', {
         filename: 'index.css',
         code: `@import 'x';`,
         warnings: [messages['only import font']]
-      },
-      {
-        title: 'prevent non custom element @kaliber-scoped in index.css',
-        filename: 'index.css',
-        code: `@kaliber-scoped abc;`,
-        warnings: [messages['only scope custom element']]
       },
     ]
   },